[WEB-4457] refactor: decouple work item properties from mobx store
