+++ apps/web/core/components/dropdowns/index.ts
@@ -1,10 +1,10 @@
-export * from "./member";
+export * from "./member/dropdown";
 export * from "./cycle";
 export * from "./date-range";
 export * from "./date";
 export * from "./estimate";
 export * from "./merged-date";
-export * from "./module";
+export * from "./module/dropdown";
 export * from "./priority";
-export * from "./project";
-export * from "./state";
+export * from "./project/dropdown";
+export * from "./state/dropdown";
