+++ apps/web/core/components/dropdowns/member/member-options.tsx
@@ -3,46 +3,48 @@
 import { useEffect, useRef, useState } from "react";
 import { Placement } from "@popperjs/core";
 import { observer } from "mobx-react";
-import { useParams } from "next/navigation";
 import { createPortal } from "react-dom";
 import { usePopper } from "react-popper";
 import { Check, Search } from "lucide-react";
 import { Combobox } from "@headlessui/react";
-import { EUserPermissions } from "@plane/constants";
+// plane imports
 import { useTranslation } from "@plane/i18n";
-// plane ui
 import { Avatar } from "@plane/ui";
 import { cn, getFileURL } from "@plane/utils";
-// helpers
 // hooks
-import { useUser, useMember } from "@/hooks/store";
+import { useUser } from "@/hooks/store";
 import { usePlatformOS } from "@/hooks/use-platform-os";
+import { IUserLite } from "@plane/types";
 
 interface Props {
-  memberIds?: string[];
   className?: string;
-  optionsClassName?: string;
-  projectId?: string;
-  referenceElement: HTMLButtonElement | null;
-  placement: Placement | undefined;
+  getUserDetails: (userId: string) => IUserLite | undefined;
   isOpen: boolean;
+  memberIds?: string[];
+  onDropdownOpen?: () => void;
+  optionsClassName?: string;
+  placement: Placement | undefined;
+  referenceElement: HTMLButtonElement | null;
 }
 
 export const MemberOptions: React.FC<Props> = observer((props: Props) => {
-  const { memberIds: propsMemberIds, projectId, referenceElement, placement, isOpen, optionsClassName = "" } = props;
+  const {
+    getUserDetails,
+    isOpen,
+    memberIds,
+    onDropdownOpen,
+    optionsClassName = "",
+    placement,
+    referenceElement,
+  } = props;
+  // refs
+  const inputRef = useRef<HTMLInputElement | null>(null);
   // states
   const [query, setQuery] = useState("");
   const [popperElement, setPopperElement] = useState<HTMLDivElement | null>(null);
-  // refs
-  const inputRef = useRef<HTMLInputElement | null>(null);
-  // store hooks
+  // plane hooks
   const { t } = useTranslation();
-  const { workspaceSlug } = useParams();
-  const {
-    getUserDetails,
-    project: { getProjectMemberIds, fetchProjectMembers, getProjectMemberDetails },
-    workspace: { workspaceMemberIds },
-  } = useMember();
+  // store hooks
   const { data: currentUser } = useUser();
   const { isMobile } = usePlatformOS();
   // popper-js init
@@ -60,21 +62,12 @@ export const MemberOptions: React.FC<Props> = observer((props: Props) => {
 
   useEffect(() => {
     if (isOpen) {
-      onOpen();
+      onDropdownOpen?.();
       if (!isMobile) {
         inputRef.current && inputRef.current.focus();
       }
     }
   }, [isOpen, isMobile]);
-
-  const memberIds = propsMemberIds
-    ? propsMemberIds
-    : projectId
-      ? getProjectMemberIds(projectId, true)
-      : workspaceMemberIds;
-  const onOpen = () => {
-    if (!memberIds && workspaceSlug && projectId) fetchProjectMembers(workspaceSlug.toString(), projectId);
-  };
 
   const searchInputKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
     if (query !== "" && e.key === "Escape") {
@@ -86,12 +79,6 @@ export const MemberOptions: React.FC<Props> = observer((props: Props) => {
   const options = memberIds
     ?.map((userId) => {
       const userDetails = getUserDetails(userId);
-      if (projectId) {
-        const role = getProjectMemberDetails(userId, projectId)?.role;
-        const isGuest = role === EUserPermissions.GUEST;
-        if (isGuest) return;
-      }
-
       return {
         value: userId,
         query: `${userDetails?.display_name} ${userDetails?.first_name} ${userDetails?.last_name}`,
