+++ apps/web/core/components/issues/select/index.ts
@@ -1,1 +1,1 @@
-export * from "./label";
+export * from "./dropdown";
